Split PCIndicator into several files.
