+++ ts/Stock/Indicators/PC/PCPoint.d.ts
@@ -1,0 +1,33 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import PCIndicator from './PCIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+*
+*  Class
+*
+* */
+
+declare class PCPoint extends SMAPoint {
+    public series: PCIndicator;
+}
+
+/* *
+*
+*  Default Export
+*
+* */
+
+export default PCPoint;
